pass compilation, buf not ready yet

+++ libraries/AC_CustomControl/util.cpp
@@ -1,11 +1,8 @@
 #include "util.h"
-// #include <iostream>
-#include <vector>
-#include <algorithm> // For std::max and std::min
 
 // Function to get the last column of a 2D vector
-std::vector<double> getLastColumn(const std::vector<std::vector<double>>& matrix) {
-    std::vector<double> last_column;
+std::vector<float> getLastColumn(const std::vector<std::vector<float>>& matrix) {
+    std::vector<float> last_column;
     for (const auto& row : matrix) {
         if (!row.empty()) {
             last_column.push_back(row.back());  // Get the last element of the row
@@ -15,32 +12,32 @@ std::vector<double> getLastColumn(const std::vector<std::vector<double>>& matrix
 }
 
 // Function to clamp elements of a vector to the range [-1, 1]
-void clampToRange(std::vector<double>& vec, double min_val = -1.0, double max_val = 1.0) {
-    for (double& element : vec) {
+void clampToRange(std::vector<float>& vec, float min_val = -1.0, float max_val = 1.0) {
+    for (float& element : vec) {
         element = std::max(min_val, std::min(element, max_val));
     }
 }
 
-std::vector<double> vecCat(const std::vector<double>& vec1, const std::vector<double>& vec2) {
-    std::vector<double> result = vec1; // Copy the first vector into the result
+std::vector<float> vecCat(const std::vector<float>& vec1, const std::vector<float>& vec2) {
+    std::vector<float> result = vec1; // Copy the first vector into the result
     result.insert(result.end(), vec2.begin(), vec2.end()); // Insert the second vector at the end of the result
     return result;
 }
 // Function to perform vector addition
-std::vector<double> vecAdd(const std::vector<double>& vec1, const std::vector<double>& vec2) {
+std::vector<float> vecAdd(const std::vector<float>& vec1, const std::vector<float>& vec2) {
     assert(vec1.size() == vec2.size());
-    std::vector<double> result(vec1.size(), 0.0);
+    std::vector<float> result(vec1.size(), 0.0);
     for (size_t i = 0; i < vec1.size(); ++i) {
         result[i] = vec1[i] + vec2[i];
     }
     return result;
 }
-std::vector<std::vector<double>> vec2DAdd(const std::vector<std::vector<double>>& c1, const std::vector<std::vector<double>>& buf) {
+std::vector<std::vector<float>> vec2DAdd(const std::vector<std::vector<float>>& c1, const std::vector<std::vector<float>>& buf) {
     // Ensure that the dimensions match
     assert(c1.size() == buf.size() && c1[0].size() == buf[0].size());
 
     // Create a result vector with the same dimensions
-    std::vector<std::vector<double>> result(c1.size(), std::vector<double>(c1[0].size(), 0.0));
+    std::vector<std::vector<float>> result(c1.size(), std::vector<float>(c1[0].size(), 0.0));
 
     for (size_t i = 0; i < c1.size(); ++i) {
         for (size_t j = 0; j < c1[i].size(); ++j) {
@@ -52,30 +49,30 @@ std::vector<std::vector<double>> vec2DAdd(const std::vector<std::vector<double>>
 }
 
 // Function to apply the ReLU activation
-std::vector<double> relu(const std::vector<double>& x) {
-    std::vector<double> result(x.size());
+std::vector<float> relu(const std::vector<float>& x) {
+    std::vector<float> result(x.size());
     for (size_t i = 0; i < x.size(); ++i) {
-        result[i] = std::max(double(0.0), x[i]);
+        result[i] = std::max(float(0.0), x[i]);
     }
     return result;
 }
 
 // Function to apply ReLU to a 2D vector
-std::vector<std::vector<double>> relu2D(const std::vector<std::vector<double>>& input) {
-    std::vector<std::vector<double>> result = input;  // Create a copy of the input
+std::vector<std::vector<float>> relu2D(const std::vector<std::vector<float>>& input) {
+    std::vector<std::vector<float>> result = input;  // Create a copy of the input
 
     for (auto& row : result) {
         for (auto& element : row) {
-            element = std::max(double(0.0), element);  // Apply ReLU
+            element = std::max(float(0.0), element);  // Apply ReLU
         }
     }
 
     return result;
 }
 // Function to apply the softmax activation
-std::vector<double> softmax(const std::vector<double>& x) {
-    std::vector<double> exp_x(x.size());
-    double sum_exp = 0.0;
+std::vector<float> softmax(const std::vector<float>& x) {
+    std::vector<float> exp_x(x.size());
+    float sum_exp = 0.0;
     for (size_t i = 0; i < x.size(); ++i) {
         exp_x[i] = std::exp(x[i]);
         sum_exp += exp_x[i];
@@ -86,12 +83,12 @@ std::vector<double> softmax(const std::vector<double>& x) {
     return exp_x;
 }
 
-std::vector<std::vector<double>> chomp1d(const std::vector<std::vector<double>>& input, int padding) {
-    std::vector<std::vector<double>> result;
+std::vector<std::vector<float>> chomp1d(const std::vector<std::vector<float>>& input, int padding) {
+    std::vector<std::vector<float>> result;
 
     for (const auto& row : input) {
         // Create a new row with the required number of elements removed from the end
-        std::vector<double> new_row(row.begin(), row.end() - padding);
+        std::vector<float> new_row(row.begin(), row.end() - padding);
         result.push_back(new_row);
     }
 
@@ -99,12 +96,12 @@ std::vector<std::vector<double>> chomp1d(const std::vector<std::vector<double>>&
 }
 
 // Helper function to apply padding
-std::vector<std::vector<double>> applyPadding(const std::vector<std::vector<double>>& input, int padding) {
+std::vector<std::vector<float>> applyPadding(const std::vector<std::vector<float>>& input, int padding) {
     int C_in = input.size();
     int L_in = input[0].size();
     int padded_L_in = L_in + 2 * padding;
 
-    std::vector<std::vector<double>> padded_input(C_in, std::vector<double>(padded_L_in, 0.0));
+    std::vector<std::vector<float>> padded_input(C_in, std::vector<float>(padded_L_in, 0.0));
 
     for (int c_in = 0; c_in < C_in; ++c_in) {
         for (int l_in = 0; l_in < L_in; ++l_in) {
@@ -115,10 +112,10 @@ std::vector<std::vector<double>> applyPadding(const std::vector<std::vector<doub
     return padded_input;
 }
 
-std::vector<std::vector<double>> conv1d(
-    const std::vector<std::vector<double>>& input,
-    const std::vector<std::vector<std::vector<double>>>& weight,
-    const std::vector<double>& bias,
+std::vector<std::vector<float>> conv1d(
+    const std::vector<std::vector<float>>& input,
+    const std::vector<std::vector<std::vector<float>>>& weight,
+    const std::vector<float>& bias,
     int stride = 1,
     int padding = 0,
     int dilation = 1) {
@@ -132,10 +129,10 @@ std::vector<std::vector<double>> conv1d(
     int L_out = (L_in + 2 * padding - dilation * (K - 1) - 1) / stride + 1;
 
     // Initialize the output tensor
-    std::vector<std::vector<double>> output(C_out, std::vector<double>(L_out, 0.0));
+    std::vector<std::vector<float>> output(C_out, std::vector<float>(L_out, 0.0));
 
     // Apply padding to the input
-    std::vector<std::vector<double>> input_padded = applyPadding(input, padding);
+    std::vector<std::vector<float>> input_padded = applyPadding(input, padding);
 
     // Perform the convolution
     for (int c_out = 0; c_out < C_out; ++c_out) {
@@ -156,9 +153,9 @@ std::vector<std::vector<double>> conv1d(
 }
 
 // Function to perform vector dot product
-double vecDot(const std::vector<double>& vec1, const std::vector<double>& vec2) {
+float vecDot(const std::vector<float>& vec1, const std::vector<float>& vec2) {
     assert(vec1.size() == vec2.size());
-    double result = 0.0;
+    float result = 0.0;
     for (size_t i = 0; i < vec1.size(); ++i) {
         result += vec1[i] * vec2[i];
     }
@@ -166,9 +163,9 @@ double vecDot(const std::vector<double>& vec1, const std::vector<double>& vec2) 
 }
 
 // Function to perform matrix-vector multiplication
-std::vector<double> matVecMul(const std::vector<std::vector<double>>& mat, const std::vector<double>& vec) {
+std::vector<float> matVecMul(const std::vector<std::vector<float>>& mat, const std::vector<float>& vec) {
     assert(mat[0].size() == vec.size());
-    std::vector<double> result(mat.size(), 0.0);
+    std::vector<float> result(mat.size(), 0.0);
     for (size_t i = 0; i < mat.size(); ++i) {
         for (size_t j = 0; j < vec.size(); ++j) {
             result[i] += mat[i][j] * vec[j];
@@ -179,9 +176,9 @@ std::vector<double> matVecMul(const std::vector<std::vector<double>>& mat, const
 
 
 // Function to apply the linear layer with activation
-std::vector<double> linear_layer(const std::vector<double>& bias, const std::vector<std::vector<double>>& weight, const std::vector<double>& input, bool activation) {
+std::vector<float> linear_layer(const std::vector<float>& bias, const std::vector<std::vector<float>>& weight, const std::vector<float>& input, bool activation) {
     // Compute the linear transformation
-    std::vector<double> output = vecAdd(bias, matVecMul(weight, input));
+    std::vector<float> output = vecAdd(bias, matVecMul(weight, input));
 
     // Apply the activation function
     if (activation) {
@@ -191,11 +188,11 @@ std::vector<double> linear_layer(const std::vector<double>& bias, const std::vec
 }
 
 // Function to apply the composition layer
-std::vector<double> composition_layer(
-    const std::vector<std::vector<std::vector<double>>>& weight, // [contextdim][outdim][indim]
-    const std::vector<std::vector<double>>& bias,               // [contextdim][outdim]
-    const std::vector<double>& comp_weight,                     // [contextdim]
-    const std::vector<double>& input,                           // [indim]
+std::vector<float> composition_layer(
+    const std::vector<std::vector<std::vector<float>>>& weight, // [contextdim][outdim][indim]
+    const std::vector<std::vector<float>>& bias,               // [contextdim][outdim]
+    const std::vector<float>& comp_weight,                     // [contextdim]
+    const std::vector<float>& input,                           // [indim]
     bool activation = true) {
     // Ensure dimensions are correct
     assert(weight.size() == bias.size());
@@ -207,13 +204,13 @@ std::vector<double> composition_layer(
     size_t outdim = weight[0].size();
     
     // Compute x = bias + weight * input for each context dimension
-    std::vector<std::vector<double>> x(contextdim, std::vector<double>(outdim));
+    std::vector<std::vector<float>> x(contextdim, std::vector<float>(outdim));
     for (size_t i = 0; i < contextdim; ++i) {
         x[i] = vecAdd(bias[i], matVecMul(weight[i], input));
     }
 
     // Compute output = comp_weight * sum(x)
-    std::vector<double> output(outdim, 0.0);
+    std::vector<float> output(outdim, 0.0);
     for (size_t i = 0; i < contextdim; ++i) {
         for (size_t j = 0; j < outdim; ++j) {
             output[j] += comp_weight[i] * x[i][j];
